fix: 'New' badge should never show to document creator, regardless of whether a view has been logged
