+++ app/components/DocumentList.js
@@ -2,7 +2,7 @@
 import ArrowKeyNavigation from "boundless-arrow-key-navigation";
 import * as React from "react";
 import Document from "models/Document";
-import DocumentPreview from "components/DocumentPreview";
+import DocumentListItem from "components/DocumentListItem";
 
 type Props = {
   documents: Document[],
@@ -18,7 +18,7 @@ export default function DocumentList({ limit, documents, ...rest }: Props) {
       defaultActiveChildIndex={0}
     >
       {items.map((document) => (
-        <DocumentPreview key={document.id} document={document} {...rest} />
+        <DocumentListItem key={document.id} document={document} {...rest} />
       ))}
     </ArrowKeyNavigation>
   );
