--- app/components/DocumentPreview/index.js
@@ -1,3 +1,0 @@
-// @flow
-import DocumentPreview from "./DocumentPreview";
-export default DocumentPreview;
