+++ app/utils/keyboard.js
@@ -5,6 +5,6 @@ export const metaDisplay = isMac ? "⌘" : "Ctrl";
 
 export const meta = isMac ? "cmd" : "ctrl";
 
-export function isMetaKey(event: KeyboardEvent) {
+export function isMetaKey(event: KeyboardEvent | MouseEvent) {
   return isMac ? event.metaKey : event.ctrlKey;
 }
